renamed tabs/tabs-list directory to tabs/list
