+++ packages/react/src/tabs/index.parts.ts
@@ -2,4 +2,4 @@ export { TabsRoot as Root } from './root/TabsRoot';
 export { TabsTab as Tab } from './tab/TabsTab';
 export { TabsIndicator as Indicator } from './indicator/TabsIndicator';
 export { TabsPanel as Panel } from './panel/TabsPanel';
-export { TabsList as List } from './tabs-list/TabsList';
+export { TabsList as List } from './list/TabsList';
